Unify frontend UI across merged branches, add persistent auth and fix model type bug
- Unified UI/CSS across all pages for consistent look and feel
- Improved visual error handling/feedback throughout the site
- Fixed model bugs where fields (e.g. DepartmentID, UserRoleID in
  User.cs) were typed as long instead of int
- Implemented Auth on all pages
- Persisted auth state in localStorage, fixing a crash that occurred
  on page reload because auth state was previously lost

+++ VagtPlan.Web/Models/UserModel.cs
@@ -6,8 +6,8 @@
         public required string Name { get; set; }
         public string? password { get; set; }
         public required int VactionDays { get; set; }
-        public required long DepartmentId { get; set; }
-        public required long UserRoleId { get; set; }
+        public required int DepartmentId { get; set; }
+        public required int UserRoleId { get; set; }
     }
 
     public class UserDto
@@ -15,8 +15,8 @@
         public required string Name { get; set; }
         public string? password { get; set; }
         public required int VactionDays { get; set; }
-        public required long DepartmentId { get; set; }
-        public required long UserRoleId { get; set; }
+        public required int DepartmentId { get; set; }
+        public required int UserRoleId { get; set; }
     }
 
     public class LoginDto
